pyproject.toml
[project]
name = "fashion-category-prediction"
version = "0.1.0"
description = "Model that recognizes fashion items"
authors = [
    {name = "Parenbz",email = "parenbz@yandex.ru"}
]
readme = "README.md"
requires-python = ">=3.11"

[tool.black]
line-length = 88
target-version = ['py36']

[tool.isort]
profile = "black"
line_length = 88

[tool.poetry.dependencies]
torch = "^2.6.0"
numpy = "^2.2.3"
torchvision = "^0.22.1"
pre-commit = "^4.1.0"
hydra-core = "^1.3"
dvc = "^3.60.0"
pandas = "^2.3.0"
tensorboard = "^2.19.0"
torcheval = "^0.0.7"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
